Update WPBridge.cs
Bumped minor version.

Added a stopwatch to measure the time from the request is sent to the response is given.

WPRequest class constructor now accepts 1 argument which holds the PlayersData list.

WPRequest class is now instantiated with the PlayersData list.

WPRequest class is now instantiated with logic to check if there are players that has left this instance. If so those players are removed from request to save bandwith.

SendPlayerData now checks to see if there are players online or if there are players who have left since last request and desides if it should send all statistics or just server statistics.

Added debug message that prints out to console how large the payload for the request is in kilobytes and how long it takes to send.

Added Plugin Hook Unload which destroys the timer object.